Изменение : починил выбор внешних субтитров, если была выставлена соответствубщая галка в настройках, так же заменил тип int на size_t, т.к. именно этот типа используется при работе с CAtlArray; Изменение : ShoutCast - доделал поддержку AAC потокового радио, не стал париться - взял рабочий код из MpegSplitter :);
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@1856 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/apps/mplayerc/MainFrm.h
@@ -346,7 +346,7 @@ class CMainFrame : public CFrameWnd, public CDropTarget
 
 	CAtlArray<SubStreams> subarray;
 	void SubFlags(CString strname, bool &forced, bool &def);
-	int GetSubSelIdx();
+	size_t GetSubSelIdx();
 	int cntintsub;
 	int SelSub2;
 	// chapters (file mode)
